Do not modify frozen eager_load_paths Array
When one does not change `config.auto_load_paths` in the
host app, by the time the `deface.tweak_eager_loading` initializer
runs, `{app,railtie}.config.eager_load_paths` is frozen [1].

Running `reject!` on a frozen Array results in an error, preventing
the host app from starting.

This commit modifies the initializer such that instead of modifying the
array, it is replaced by the modified array, preventing this error.

[https://github.com/rails/rails/blob/master/railties/lib/rails/engine.rb#L573]